web_app/backend/requirements.txt
numpy==1.25.2
flask==2.3.3
flask-cors==4.0.0
faiss-cpu==1.11.0
sentence-transformers==2.2.2
torch==2.0.1
transformers==4.31.0
huggingface-hub==0.16.4
openai==1.93.0
python-dotenv==1.1.1
requests==2.31.0
torchvision
torchaudio
gunicorn
boto3==1.34.69 